Move end to end tests into separate folder
